requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
httpx>=0.27.0
python-dotenv>=1.0.0
playwright>=1.40.0
# After installing playwright, run: playwright install chromium

# Persistence
sqlalchemy>=2.0.0

# Rate limiting
slowapi>=0.1.9

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
